OCD-887: addresses all code review comments except Maven
