+++ .gitignore
@@ -17,6 +17,8 @@ local.properties
 .settings/
 .loadpath
 target/
+fullResults.txt
+products.csv
 
 # External tool builders
 .externalToolBuilders/
@@ -212,46 +214,3 @@ pip-log.txt
 
 #Translations
 *.mo
-
-#Mr Developer
-.mr.developer.cfg
-
-## CHPL specific ignores
-/chpl/chpl-etl/src/main/resources/log.txt
-/chpl/chpl-etl/src/main/resources/chpl.csv
-/chpl/chpl-etl/src/main/resources/chpl-hash.csv
-/chpl/chpl-etl/src/main/resources/chpl-raw.csv
-/chpl/chpl-etl/src/main/resources/chpl-small.xlsx
-/chpl/chpl-etl/src/main/resources/chpl-wChecksum.csv
-/chpl/chpl-etl/src/main/resources/chpl_to-insert.csv
-/chpl/chpl-etl/src/main/resources/chpl_to-update.csv
-/chpl/chpl-etl/src/main/resources/log-create_certified-product.txt
-/chpl/chpl-etl/src/main/resources/log_checksum-analysis.txt
-/chpl/chpl-etl/src/main/resources/log_create_vendor-product.txt
-/chpl/chpl-etl/src/main/resources/log_unchanged-rows.csv
-/chpl/chpl-etl/parameters.prm
-/chpl/chpl-etl/src/main/resources/log-uncreated_certified.txt
-/chpl/chpl-etl/src/main/resources/log-unmerged_certified.txt
-/chpl/chpl-etl/src/main/resources/log-uncreated_products.txt
-/chpl/chpl-etl/chpl_to-update.csv
-/chpl/chpl-etl/log-uncreated_certified.txt
-/chpl/chpl-etl/log-unmerged_certified.txt
-/chpl/chpl-etl/log_checksum-analysis.txt
-/chpl/chpl-etl/log_unchanged-rows.csv
-/chpl/chpl-etl/log_create-vendor-product.txt
-/chpl/chpl-etl/log_create_certified-product.txt
-/chpl/chpl-etl/log_uncreated-certified.txt
-/chpl/chpl-etl/log_unmerged-certified.txt
-/chpl/chpl-etl/to-update.csv
-/chpl/chpl-etl/cert_products_x_ids_null.txt
-/chpl/chpl-etl/input/
-/chpl/chpl-etl/parse.sh
-/chpl/chpl-etl/log_checksum-analysis-finallines.txt
-/openchpl-sql/openchpl-role.sql
-/chpl/chpl-service/src/main/resources/environment.properties
-/chpl/chpl-service/src/main/resources-dev/environment.properties
-/chpl/chpl-service/src/main/resources-staging/environment.properties
-/chpl/chpl-service/src/main/resources-production/environment.properties
-/chpl/chpl-service/src/test/resources/environment.test.properties
-/chpl/chpl-auth/src/test/resources/environment.auth.test.properties
-/chpl/chpl-service/test-target/
