--- index-files/index-3.html
@@ -1,129 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<!-- NewPage -->
-<html lang="en">
-<head>
-<!-- Generated by javadoc (1.8.0_92) on Thu Nov 10 16:57:55 EST 2016 -->
-<title>L-Index</title>
-<meta name="date" content="2016-11-10">
-<link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
-<script type="text/javascript" src="../script.js"></script>
-</head>
-<body>
-<script type="text/javascript"><!--
-    try {
-        if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="L-Index";
-        }
-    }
-    catch(err) {
-    }
-//-->
-</script>
-<noscript>
-<div>JavaScript is disabled on your browser.</div>
-</noscript>
-<!-- ========= START OF TOP NAVBAR ======= -->
-<div class="topNav"><a name="navbar.top">
-<!--   -->
-</a>
-<div class="skipNav"><a href="#skip.navbar.top" title="Skip navigation links">Skip navigation links</a></div>
-<a name="navbar.top.firstrow">
-<!--   -->
-</a>
-<ul class="navList" title="Navigation">
-<li><a href="../package-summary.html">Package</a></li>
-<li>Class</li>
-<li>Use</li>
-<li><a href="../overview-tree.html">Tree</a></li>
-<li><a href="../deprecated-list.html">Deprecated</a></li>
-<li class="navBarCell1Rev">Index</li>
-<li><a href="../help-doc.html">Help</a></li>
-</ul>
-</div>
-<div class="subNav">
-<ul class="navList">
-<li><a href="index-2.html">Prev Letter</a></li>
-<li><a href="index-4.html">Next Letter</a></li>
-</ul>
-<ul class="navList">
-<li><a href="../index.html?index-files/index-3.html" target="_top">Frames</a></li>
-<li><a href="index-3.html" target="_top">No&nbsp;Frames</a></li>
-</ul>
-<ul class="navList" id="allclasses_navbar_top">
-<li><a href="../allclasses-noframe.html">All&nbsp;Classes</a></li>
-</ul>
-<div>
-<script type="text/javascript"><!--
-  allClassesLink = document.getElementById("allclasses_navbar_top");
-  if(window==top) {
-    allClassesLink.style.display = "block";
-  }
-  else {
-    allClassesLink.style.display = "none";
-  }
-  //-->
-</script>
-</div>
-<a name="skip.navbar.top">
-<!--   -->
-</a></div>
-<!-- ========= END OF TOP NAVBAR ========= -->
-<div class="contentContainer"><a href="index-1.html">D</a>&nbsp;<a href="index-2.html">G</a>&nbsp;<a href="index-3.html">L</a>&nbsp;<a href="index-4.html">M</a>&nbsp;<a href="index-5.html">S</a>&nbsp;<a name="I:L">
-<!--   -->
-</a>
-<h2 class="title">L</h2>
-<dl>
-<dt><span class="memberNameLink"><a href="../SampleApp.html#loadProperties-java.lang.String-">loadProperties(String)</a></span> - Method in class <a href="../SampleApp.html" title="class in &lt;Unnamed&gt;">SampleApp</a></dt>
-<dd>
-<div class="block">Utility method to load Properties from environment.properties file.</div>
-</dd>
-</dl>
-<a href="index-1.html">D</a>&nbsp;<a href="index-2.html">G</a>&nbsp;<a href="index-3.html">L</a>&nbsp;<a href="index-4.html">M</a>&nbsp;<a href="index-5.html">S</a>&nbsp;</div>
-<!-- ======= START OF BOTTOM NAVBAR ====== -->
-<div class="bottomNav"><a name="navbar.bottom">
-<!--   -->
-</a>
-<div class="skipNav"><a href="#skip.navbar.bottom" title="Skip navigation links">Skip navigation links</a></div>
-<a name="navbar.bottom.firstrow">
-<!--   -->
-</a>
-<ul class="navList" title="Navigation">
-<li><a href="../package-summary.html">Package</a></li>
-<li>Class</li>
-<li>Use</li>
-<li><a href="../overview-tree.html">Tree</a></li>
-<li><a href="../deprecated-list.html">Deprecated</a></li>
-<li class="navBarCell1Rev">Index</li>
-<li><a href="../help-doc.html">Help</a></li>
-</ul>
-</div>
-<div class="subNav">
-<ul class="navList">
-<li><a href="index-2.html">Prev Letter</a></li>
-<li><a href="index-4.html">Next Letter</a></li>
-</ul>
-<ul class="navList">
-<li><a href="../index.html?index-files/index-3.html" target="_top">Frames</a></li>
-<li><a href="index-3.html" target="_top">No&nbsp;Frames</a></li>
-</ul>
-<ul class="navList" id="allclasses_navbar_bottom">
-<li><a href="../allclasses-noframe.html">All&nbsp;Classes</a></li>
-</ul>
-<div>
-<script type="text/javascript"><!--
-  allClassesLink = document.getElementById("allclasses_navbar_bottom");
-  if(window==top) {
-    allClassesLink.style.display = "block";
-  }
-  else {
-    allClassesLink.style.display = "none";
-  }
-  //-->
-</script>
-</div>
-<a name="skip.navbar.bottom">
-<!--   -->
-</a></div>
-<!-- ======== END OF BOTTOM NAVBAR ======= -->
-</body>
-</html>
