--- products.csv
@@ -1,21 +1,0 @@
-versionId,name,id,version
-3981,1 Connect BuildYourEMR,1736,5.0
-5234,Access Profiler Enterprise,626,V6.1
-302,Accelechart,71,1.2
-5234,Access Profiler Enterprise,626,V6.1
-934,4medica iEHR Cloud Ambulatory Suite,29,15.10
-784,AbbaDox® EHR,1075,14
-5978,ABELMed EHR - EMR / PM,2581,12
-5978,ABELMed EHR - EMR / PM,2581,12
-5052,abcEHR,1466,MU 1.0
-2820,2013 Systemedx Clinical Navigator,2246,2013.12
-3647,365EHR,594,4.0.14
-5234,Access Profiler Enterprise,626,V6.1
-4929,A.I. Med Pro,76,9.2.0
-5975,A.I.med Pro,75,version 9.3
-5158,37 Degrees,2215,V1.07
-5978,ABELMed EHR - EMR / PM,2581,12
-5234,Access Profiler Enterprise,626,V6.1
-4317,24/7 smartEMR,2379,6.0
-5978,ABELMed EHR - EMR / PM,2581,12
-4921,A.I.med Pro,78,9.2
